Fix a test helper to better handle the possibility of a short read
The TestClient_Chroot test was seen to fail once from what looks like a short
read. The client tests are basically reading off of a socket, so not very
complicated, but the test may have had a short read where the read returned,
but had no content. In this case, it will retry again for up to 5 times,
however return once it has any kind of content.

Signed-off-by: Ken Robertson <ken@apcera.com>