Updates to appease gcc 5.1
My local Arch machine updated to GCC 5.1, which looks like it has a change to
default to enable warning about implicit declarations.

The code still compiled, but was getting a slew of warnings from stage2/stage3
since we had some functions not defined in the headers, and some functions
they were using that weren't explicitly included.