Replace use of pivot_root with mount and move.
This changes the way the filesystem is scoped from using pivot_root to instead
bind mount the directory over itself, and then move it to the root. This
approach is simpler and less black magic than pivot_root, and also resolves
the issue we've seen before where multiple contains calling pivot_root at the
same time is racey.

This also changes how host filesystem access is setup in privileged mode.
Previously, it just left the old host filesystem mounted with pivot_root.

Now, have added a new MOUNT command to the stage3 init process. With this, the
stage1 can ask the stage3 to perform the bind mounts before it performs the
chroot.

It will now set up /host/pods going directly to the pods filesystem (read-
only). It will also set up /host/proc to be the proc partition of the host.
Note the comment regarding it overall not being read-only, mostly due to
limitations in fully enforcing read-only through things like traversing
/proc/pid/root to processes root not maintaining read-only state.