Change direct API from TCP to unix socket based.
This changes the core Kurma process to use a unix socket for all API access.
This allows the API to be isolated rather than having to rely on notions of
authentication.

Locally, a Kurma process should only be launching very trusted containers and
giving them host API access. Or, the process launching a container for host API
access itself has API access, so Kurma is trusting it.

Generally, if a container process has API access, it should be trusted. This is
one of the base ideas of Kurma as a base building block, and with some initial
trusted containers layering on logic about a broader cluster, and it defining
what the local host should trust or not.

The kurma-api process has been updated to proxy to the local host unix socket
from its TCP socket. It also rejects any container requesting host API access.

The kurma-cli has been updated to default to using the unix socket.