client/stage1: Add support for the "enter" command to be passed in the command
This updates the "enter" API call to allow it to pass in the command that
should be executed.

Previously, the command was hard coded /bin/bash, however this was a bad
assumption as some containers may simply be busybox based and have /bin/sh.

Additionally, there are plenty of cases where the user may with to execute
something specific other than a shell, so the API has been updated to allow
the command to be provided.

If no command is provided, the base assumption has been changed to use /bin/sh.