Fixed instrumentation of udev service.
When a udev container was being used, it was failing to setup due to disk
mounts happening over it for the pod container space. Instead, have set up a
side directory for "system containers", where it is a container setup before
all disk configuration happens.

I'm not particularly fond of how it is configured. It is swapping out the
manager's containerDirectory while setting up. Want to change that in the
future where a container can specify some values to override the manager's.