Added support for A/B boot partitions and upgrade rollback support.
This adds support for updating the GPT settings on a successful boot. It can
be used to mark a new kernel as ready for a test boot on an upgrade, then
setting it as successful once booted. If it fails to boot it, then the tries
field will be at 0, and a reboot will trigger the host to boot up on the
previous kernel image.

This also adds utils/installer, which is used to update a KurmaOS host to use
a new kernel and then trigger it to be booted using kexec.