Add switch_root functionality directly to Kurma's bootstrap
This adds handling for moving off of the initramfs and onto a tmpfs root
directly to Kurma's bootstrap. Previously, this was handled by an busybox
script on boot, and was somewhat hackish using cp to copy the files and getting
recursion warnings.

This reduces the use of executables in the init image and moves more setup
internal, and doing it more cleanly.