Fix issue with pointer when evaluating boot parameters.
This fixes an issue where the pointer from the variable when looping
boot parameters was assigned to the configuration. If that wasn't the
last element iterated in the map, then the point would be changed to
point to another argument later.

This ensures the variable is assigned in the local scope and then it
takes the pointer of that.