Adds support for resource and capability isolators.
This adds support for the CPU and memory resource isolators through
cgroups.

This also adds support for Linux capabilities isolators, to either
specify the retained or the revoked set of capabilities. Capabilities
will be applied within the stage3 process on START requests. It won't
apply it to the stage3 process itself, as the scoped capabilities may
prevent it from the starting the application.

This also adds support for capabilities to the stage2 process. This is
used with the enter command, and ensures entering the container
preserves the intended capabilities.