Minor fixes and cleanup.
Removed a place where a debug flag was being set for stage3 from testing
the API isolator support.

Fixed an issue where volume info was failing to be applied to the pod.

Fixed an issue where the stage2 process was not properly breaking in a
switch statement. Confusion between Go and C default behavior.

Fixed a stage3 wait request test that was seen to be racey during
development. Ensure it waits on the waitgroup before checking the
flag. Also removed the hard coded sleep in the test and use a timeout
instead.