Change format of the linux namespace isolator.
Previously, the isolator was an array of namespaces the container should
have as its own. However, this was somewhat confusing, as in most cases,
the image should be opting out of specific namespaces, rather than
specifying the set it wants.

Additionally, as an array, the range of values was limiting. The goal of
this change is to make it more explicit which namespaces should be
shared with the host, but also so that the format could in the future
support sharing a namespace from another container. With this, the value
could be the UUID of another container it wishes to share the namespace
with.