Fix issue with bad interaction in pid and user namespace creation.
This fixes an issue where the pid namespace was being created via an
unshare call before the user namespace was being created. Part of
creating the user namespace is to fork and use the child to populate the
uid/gid maps.

This fixes it to ensure the unshare call for the pid namespace happens
right before the fork call intended to create pid 1.