stage2: Ensure the root mount is changed to private in the new namespace
Ensure that the host's root filesystem is changed to a private mount
within the new mount namespace before configuring the root.

It is possible for the initial mount to /tmp can propagate to the parent
namespace, overriding its /tmp directory. By converting the root to
private within the new namespace, this better ensures the future changes
are isolated to its namespace.