Moved making root mount private to be done earlier.
Moved remounting the host's root as private within the new mount
namespace to be done earlier, after the new mount namespace is created.

With it only in createroot(), it wouldn't run if we were creating a
mount namespace, but not creating the root, such as when the
graphstorage modules run to create the overlay mount. In this case,
it'll propagate to the host.

By doing it earlier, it ensures anytime a new mount namespace is created,
it is done.