init: Support preloading embedded ACI images
With preloading the embedded images, this allows the embedded images to
have dependencies, as long as their fulfilled with the ones
provided. This is primarily to move where the console depends on a
busybox image, and multiple things end up using the busybox image.

This is currently lazy about handling lookup when no version is
specified. Some expansion needs to be done in the image manager about
multiple versions pointing to the same manifest, such as having an image
with a version string also be considered latest.