cli: Allow enter command to work when input is not a terminal
This changes the `enter` command to function when the input is not a
terminal. Changing to raw mode will no longer result in an error, and
when reading from stdin finishes, it will ensure to write
end-of-transmission to the websocket.

This allows the command to take input from a file or pipe.