Adds network plugin support.
This adds network plugin functionality to Kurma, where it can be
configured with a set of containers to launch to handle networking
instrumentation for other containers.

This is an initial implementation mostly intended for kurma-init
mode. Some API to manage the network plugins is still needed.

See Documentation/network_specification.md for the full plugin
specification.

This also includes updating of a number of dependencies.

Also includes a refactoring of the image manager and container manager,
where now they are implementing an interface which allows easier
handling with inter dependencies and should be better for more isolated
unit testing.

Fixes #25.