Fix TestCgroup_AddTask around newer cgroups handling and Docker running
This updates the TestCgroup_AddTask test to properly check when the
cgroup type is combined with "cpu,cpuacct", while before they would be
two separate entries.

This also changes it to validate against the suffix of the cgroup, since
when the tests are running within Docker, it will include the path of
the current Docker container.