Update stage3 tests to better handle read conditions.
This adds more retries to the ReadAll call, returns if it reaches an
EOF, and a small pause in between tries.

This should better handle some occassional bad reads. An EOF should be
reached when the connection is actually closed. All the individual tests
use donech on a select with a timeout, so it won't go on indefinitely in
any test.