Adds support to dynamically convert docker images to ACI.
This adds support to aciremote to take in image URLs in the form of
"docker://nats" to signify a Docker registry image. This will use
docker2aci to convert the image into an ACI and return the os.File
pointing to it.

This is used now by kurma-init for the initial containers as well as by
kurma-cli, so "kurma-cli create" can now be used to deploy local files,
ACI discovery images, and Docker registry images.