testing: Adds integration tests for kurmad
This adds an integration testing framework for testing against kurmad
using kurma-cli and the API.

The framework uses rspec to allow instrumenting tests to be simple and
having a rich matching library and ecosystem of gems to pull in to help
write the tests.

Currently, there are just a few simplistic tests, but will grow the
tests over time.