Migrate vendor management to bmizerany/vendor
After some struggles with both glide and godep, figure simple is better.

Glide wanted to use submodules, so doing any updates would have it clear
SHA fields or use the current repo's. It also bundled a lot of non-code
things like images/svgs.

Godep is a struggle with movement in dependencies of dependencies. It'll
flatten out the tree then views child dependencies as your depndencies,
and when a dependency changes its dependencies, it becomes more
complicated than it should be.

Simple 'vendor' tool from Blake is plain and simple.

This change includes updating to the latest of some dependencies.