Major refactor to introduce staging and pod functionality
This represents a major refactoring of the codebase. Some of the changes
include:

* stagers

Stagers are a means of doing pluggable workload orchestration. This
moves away from the baked in notion Kurma had before with its
structure. This reduces the instrumentation and management done by the
top level kurmad and pushes logic down to a separate process responsible
for managing an individual workload. For more information, see the
specification in the Documentation directory.

* pods

Aside from a massive rename of "container" to "pod", the first stager
implementation is one that focuses on being able to instrument multiple
apps (or containers) within a single pod, where they share common
namespaces and have their own mount namespace.

This also includes changes in how workloads are launched, focusing on
providing a pod manifest to define a set of applications.

* libcontainer

Our previous spawner/initd implementations have been removed in favor of
transitioning container instrumentation to use
libcontainer. libcontainer makes a lot of sense here as it is well used
and vetted, works perfectly in the library context, and offers a lot of
functionality built in that Kurma hasn't yet needed, but will in the
future.

* networking updates

Network management has been updated to focus on instrumenting a single
networking pod which contains all of the network plugins as individual
applications. This centralizes their instrumentation more.

Networking plugins no longer share the PID namespace with the
host. Instead, it focuses on a volume being attached to the plugin which
contains a directory where networking namespaces are bind mounted in.

Part of this is with the changes around workload setup from stagers, the
network namespace is setup a head of the stager being launched so
networking is already configured for the stager. This way, the song and
dance coordination between kurmad and the stager can be more straight
forward rather than "you do this, then I do that".

* code restructuring

The codebase has been largely restructure to better represent some of
the functional areas. Previously, the organization was not very
logically composed with the stage1/2/3, especially as stages 2/3 were
removed.

* interface definitions

Interfaces have been defined for some of the main functional areas like
PodManager, ImageManager, NetworkManager. This has two benefits: clean
up the import tree by having all packages import a central definition
and code to the interface, rather than spaghetti and risk circular
references, and also, tests can easily scope to the individual component
with mocked implementations of the interfaces it might reference.

* kurmad configuration

kurmad has been updated to have configuration file handling for
specifing path configuration, network definitions, and initial pods to
run on start. Previously, kurmad was lagging behind kurmaOS for these
definitions. Going forward, kurmad will be kept in parity.

* testing

Have begun back filling a lot of packages that were light on test
coverage. This is a lot easier following the interface definitions, and
more tests will be getting filled in over time.

* unique pod name enforcement

Unique pod names are now enforced. This was a ticket that was open on
Github and was quickly addressed during some of the work.