Updates for developer documentation.
Added 'make help' which will print out a short help message.

Added 'make run' which will launch kurma-server and allow you to connect
up to it. On Linux, it will run it locally and print the path to the
socket. On other, it'll run Docker and print how to connect to the api
proxy.

Added some updates to the readme with documentation on developer setup.

Update the api-proxy ACI to run within the local dev mode.

Update kurmad setup to handle relative paths in the configuration for
directory entries (pods, volumes, images).