stager: Restructured the container stager to a single executable
Restructured the container stager to focus on each of the call in points
being a package and a single executable main function.

With this, each package will be able to test individually rather than
multiple main() functions everywhere.

Additionally, with a single executable that switches off of os.Args[0],
the size of the stager's ACI image is reduced by about 35%.