kurmad: Add support for signal handling to do clean shutdown
This adds signal handling for SIGTERM, SIGINT, and SIGQUIT to perform a
graceful shutdown, including tearing down all of the running pods.

In the future, plan on instrumenting restarting without tearing down
pods (likely with a SIGUSR1) and reloading configuration (likely
SIGHUP).

This adds a Shutdown method to the PodManager interface. The
implementation makes sure to leave the networking pod for last, so that
all other pods will be able to have their networking deprovisioned
properly.

Also changed the integration tests to use SIGTERM rather than
SIGQUIT. It generally is a better fit.