stager: Update run command to properly handle ttys
This updates the stager specification and container implementation
around the tty handling on the run command. The run command is no longer
sent a normal AppC app object, instead it is now a slightly different
type. It formally removes some fields that were not handled before like
MountPoints and Isolators, but also adds a boolean field for Tty which
controls whether the process that is called should be given a tty.

The previous instrumentation to allocate a pty was removed from the
`pkg/daemon` handling, instead it just passes through the `net.Conn`
from the websocket.

In the stager is the best location, since need to also ensure `setctty`
is used when launching the end command. Previously, the pty was making
it down to the process, but wasn't configured correctly after the pod
refactor. The old stage2 would check if the input was a terminal and set
`Setctty`, however need to go through a different mechanism with
libcontainer.

This removes the `github.com/kr/pty` dependency as it is no longer used.