Add support for populating the resolv.conf in pods
This adds support for populating the resolv.conf in the podmanager
when it is setting up the stager. It will look for a returned DNS
configuration from the networkmanager and use that. If none is found,
then it will parse the host's resolv.conf and use that to write a
resolv.conf for the stager. It does this rather than mount in the
resolv.conf to ensure it is scoped to the necessary settings.

The stager has also been updated to simply bind mount its resolv.conf
into the individual applications. It is mounted as read only so apps
cannot alter it.

Unit tests added around the podmanager's setup for the
stager. Integration test added to ensure it works all the way down
through to the app containers.

A small change was included to allow `kurma-cli create` to take
additional arguments which will override the package's default Exec
command.