stager/container: Ensure output is written before exiting
This fixes an issue where the 'run' callin to the container stager could
exit when the child process is done, but not before the goroutines to
copy data over the tty are finished writing.

This adds a waitgroup around copying from the tty to stdout and ensures
it has finished before the process exits.

Additionally, added a simple integration test that will enter a
container and simply run `whoami`. This is just a simpler test to
validate the 'enter' command than testing the resolv.conf writing.