podmanager: Ensure cgroups mounts are private
On some platforms, such as Ubuntu Xenial, the cgroups location defaults
to being shared. The mapping in of cgroups can then cause an exponential
growth of the mountinfo file which causes large CPU spikes parsing it on
container spin up.

This ensures the cgroups mounts that are mapped in are private and do
not propagate. This is controlled just at the launching of the stager by
the podmanager.