Add support for aufs to graphstorage.
This adds support for aufs as a graphstorage implementation. This
enables Kurma to work on Ubuntu 14.04 and 12.04.

The container stager has also been updated to default to automatically
detecting which driver to use, if one isn't explicitly defined.

A few updates have been made to the Ubuntu packer definitions for Ubuntu
14.04 and 12.04. On both, the cgroup-lite package was installed to
ensure all cgroups mounts are available on boot. Additionally, the 12.04
defintion will install the 3.13 kernel from Ubuntu trusty. By default,
12.04 used the 3.2 kernel, which is older than kernels typically
recommended for container workloads for full and stable functionality.

Fixes #33.