makefile: Updated release handling to build tarballs
This updates the release make tasks to generate releasable
tarballs. These include the license file as well as all the necessary
dependencies and example configuration, such as for the main kurmad
release.

Previously, the release just generated the set of binaries just losely
in a directory, but not a single file that could be downloaded,
extracted, and utilized.