podmanager: Fix an issue with resolv.conf when using an isolated network
The CNI modules will default to returning a result including `"dns:"{}`,
which will cause the DNS property on the result ot be populated with an
empty object rather than nil.

This adds a work around which iterates over all of the DNS object's
fields and if all of them are blank/empty, it will use the host's
resolv.conf settings.

This changes some of the system tests so that the resolv.conf check is
done with both host networking and isolated networking, currently using
the default included bridge network. In time, will likely have multiple
networks defined in the testing config for testing more networking
modes/settings.