stager/api: Add support for returning the exit code on run calls
This updates the container stager to properly return the exit code from
the process it was executing. Previously, the caller was always getting
a zero exit code rather than the actual one.

This also adds support for returning the exit code on container enter
requests over the API/CLI. This way, the CLI will exit with the same
code as the process it was running in the container.

An integration test has been added to test the exit code handling
propagating.