packer: Specify delete_on_termination in packer templates
This adds explicitly setting delete_on_termination to all the packer
templates. Without this specified, it was copying the setting from the
source image. On all of the source images except for CentOS, this was
enabled by default.

However, without it enabled on CentOS, CI integration test runs were
leaving behind volumes on AWS for each run.

Explicitly setting this will ensure it is always enabled.